removed some warnings, fixed protocol, added bytes parser

+++ app/src/main/java/com/example/kestutis/cargauges/holders/RealTimeDataHolder.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RealTimeDataHolder {
-    private int _oilTemperature;
-    private int _oilPressure;
-    private int _waterTemperature;
-    private int _charge;
+    private float _oilTemperature;
+    private float _oilPressure;
+    private float _waterTemperature;
+    private float _charge;
 }
